+++ quadratic-core/src/bin/export_types.rs
@@ -55,8 +55,8 @@ fn main() {
         // Value,
         Instant,
         Duration,
-        wasm_bindings::JsFormulaParseResult,
-        wasm_bindings::JsCellRefSpan,
+        wasm_bindings::FormulaParseResult,
+        wasm_bindings::CellRefSpan,
         Error,
         ErrorMsg,
         Pos,
